refactor: 将能共同使用的方法都放入了 base brush 中
