+++ packages/s2-core/src/interaction/data-cell-brush-selection.ts
@@ -7,8 +7,6 @@ import {
   InteractionStateName,
 } from '../common/constant/interaction';
 import type { BrushRange, ViewMeta } from '../common/interface';
-import { getCellMeta } from '../utils/interaction/select-event';
-import { getActiveCellsTooltipData } from '../utils/tooltip';
 import { BaseBrushSelection } from './base-brush-selection';
 
 /**
@@ -21,42 +19,9 @@ export class DataCellBrushSelection extends BaseBrushSelection {
 
   protected bindMouseDown() {
     this.spreadsheet.on(S2Event.DATA_CELL_MOUSE_DOWN, (event: CanvasEvent) => {
-      event?.preventDefault?.();
-      if (this.spreadsheet.interaction.hasIntercepts([InterceptType.CLICK])) {
-        return;
-      }
-      this.setBrushSelectionStage(InteractionBrushSelectionStage.CLICK);
-      this.initPrepareSelectMaskShape();
-      this.setDisplayedCells();
-      this.startBrushPoint = this.getBrushPoint(event);
-      this.resetScrollDelta();
+      super.mouseDown(event);
     });
   }
-
-  protected renderPrepareSelected = (point: Point) => {
-    const { x, y } = point;
-    const target = this.spreadsheet.container.getShape(x, y);
-
-    const cell = this.spreadsheet.getCell(target);
-
-    if (!cell || !(cell instanceof DataCell)) {
-      return;
-    }
-    const { rowIndex, colIndex } = cell.getMeta();
-
-    this.endBrushPoint = {
-      x,
-      y,
-      rowIndex,
-      colIndex,
-    };
-
-    const { interaction } = this.spreadsheet;
-    interaction.addIntercepts([InterceptType.HOVER]);
-    interaction.clearStyleIndependent();
-    this.updatePrepareSelectMask();
-    this.showPrepareSelectedCells();
-  };
 
   protected bindMouseMove() {
     this.spreadsheet.on(S2Event.GLOBAL_MOUSE_MOVE, (event) => {
@@ -81,48 +46,6 @@ export class DataCellBrushSelection extends BaseBrushSelection {
     });
   }
 
-  protected bindMouseUp() {
-    // 使用全局的 mouseup, 而不是 canvas 的 mouse up 防止刷选过程中移出表格区域时无法响应事件
-    this.spreadsheet.on(S2Event.GLOBAL_MOUSE_UP, (event) => {
-      if (this.brushSelectionStage !== InteractionBrushSelectionStage.DRAGGED) {
-        this.resetDrag();
-        return;
-      }
-      this.clearAutoScroll();
-
-      // console.log(this.isValidBrushSelection(), 'is valid')
-      if (this.isValidBrushSelection()) {
-        this.spreadsheet.interaction.addIntercepts([
-          InterceptType.BRUSH_SELECTION,
-        ]);
-        this.updateSelectedCells();
-        this.spreadsheet.showTooltipWithInfo(
-          event,
-          getActiveCellsTooltipData(this.spreadsheet),
-        );
-      }
-      if (
-        this.spreadsheet.interaction.getCurrentStateName() ===
-        InteractionStateName.PREPARE_SELECT
-      ) {
-        this.spreadsheet.interaction.reset();
-      }
-
-      this.resetDrag();
-    });
-
-    // 刷选过程中右键弹出系统菜单时, 应该重置刷选, 防止系统菜单关闭后 mouse up 未相应依然是刷选状态
-    this.spreadsheet.on(S2Event.GLOBAL_CONTEXT_MENU, () => {
-      if (
-        this.brushSelectionStage === InteractionBrushSelectionStage.UN_DRAGGED
-      ) {
-        return;
-      }
-      this.spreadsheet.interaction.removeIntercepts([InterceptType.HOVER]);
-      this.resetDrag();
-    });
-  }
-
   protected isInBrushRange(meta: ViewMeta) {
     const { start, end } = this.getBrushRange();
     const { rowIndex, colIndex } = meta;
@@ -133,29 +56,6 @@ export class DataCellBrushSelection extends BaseBrushSelection {
       colIndex <= end.colIndex
     );
   }
-
-  // 获取对角线的两个坐标, 得到对应矩阵并且有数据的单元格
-  protected getBrushRangeDataCells(): DataCell[] {
-    this.setDisplayedCells();
-    return this.displayedCells.filter((cell) => {
-      const meta = cell.getMeta();
-      return this.isInBrushRange(meta);
-    });
-  }
-
-  // 刷选过程中高亮的cell
-  protected showPrepareSelectedCells = () => {
-    const brushRangeDataCells = this.getBrushRangeDataCells();
-    this.spreadsheet.interaction.changeState({
-      cells: brushRangeDataCells.map((item) => getCellMeta(item)),
-      stateName: InteractionStateName.PREPARE_SELECT,
-      // 刷选首先会经过 hover => mousedown => mousemove, hover时会将当前行全部高亮 (row cell + data cell)
-      // 如果是有效刷选, 更新时会重新渲染, hover 高亮的格子 会正常重置
-      // 如果是无效刷选(全部都是没数据的格子), brushRangeDataCells = [], 更新时会跳过, 需要强制重置 hover 高亮
-      force: true,
-    });
-    this.brushRangeCells = brushRangeDataCells;
-  };
 
   public getSelectedCellMetas = (range: BrushRange) => {
     const metas = [];
@@ -188,7 +88,7 @@ export class DataCellBrushSelection extends BaseBrushSelection {
   };
 
   // 最终刷选的cell
-  private updateSelectedCells() {
+  protected updateSelectedCells() {
     const { interaction } = this.spreadsheet;
 
     const range = this.getBrushRange();
